refactor: semver -> npm
