+++ lib/versioning/hashicorp/index.js
@@ -1,4 +1,4 @@
-const semver = require('../semver');
+const semver = require('../npm');
 
 function hashicorp2npm(input) {
   // The only case incompatible with semver is a "short" ~>, e.g. ~> 1.2
