+++ lib/versioning/npm/index.js
